LOOPCARE-745 add food item in the recipe and log recipe to the meal

See merge request mobidev/loopcare-frontend!158

+++ lib/features/nutrition/application/recipe/dto/add_food_item_to_recipe_body.dart
@@ -9,7 +9,7 @@ abstract class AddFoodItemToRecipeBody implements _$AddFoodItemToRecipeBody {
   const AddFoodItemToRecipeBody._();
 
   const factory AddFoodItemToRecipeBody({
-    required int numberOfUnits,
+    required double numberOfUnits,
     required String servingId,
   }) = _AddFoodItemToRecipeBody;
